# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Back-end Challenge - Dictionary",
        "version": "1.0"
    },
    "paths": {
...
        "/entries/en/{word}": {
            "get": {
                "tags": [
                    "Words"
                ],
                "summary": "Get details of a specific word",
                "operationId": "5b0b5ee5fad7f1a498fece48906b450c",
                "parameters": [
                    {
                        "name": "word",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Details of the word",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Word"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Word not found"
                    }
                }
            }
        },
        "/entries/en/{word}/favorite": {
            "post": {
                "tags": [
                    "Words"
                ],
                "summary": "Favorite a word",
                "operationId": "80802e491a2fe14c7f0ca63ca62b856d",
                "parameters": [
                    {
                        "name": "word",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Word not found"
                    }
                }
            }
        },
        "/entries/en/{word}/unfavorite": {
            "delete": {
                "tags": [
                    "Words"
                ],
                "summary": "Unfavorite a word",
                "operationId": "e8a40d2807ece60bde3dd56715030dea",
                "parameters": [
                    {
                        "name": "word",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Word not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Definition": {
                "title": "Definition",
                "description": "Definition model",
                "properties": {
                    "id": {
                        "description": "ID of the definition",
                        "type": "integer"
                    },
                    "definition": {
                        "description": "Text of the definition",
                        "type": "string"
                    },
                    "example": {
                        "description": "Example usage of the word in the definition",
                        "type": "string"
                    },
                    "meaning_id": {
                        "description": "ID of the related meaning",
                        "type": "integer"
                    },
                    "meaning": {
                        "$ref": "#/components/schemas/Meaning"
                    },
                    "synonyms": {
                        "description": "List of synonyms for the definition",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DefinitionSynonym"
                        }
                    },
                    "antonyms": {
                        "description": "List of antonyms for the definition",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DefinitionAntonym"
                        }
                    },
                    "created_at": {
                        "description": "Creation timestamp of the definition",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp of the definition",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "DefinitionAntonym": {
                "title": "DefinitionAntonym",
                "description": "Antonym for the definition",
                "properties": {
                    "id": {
                        "description": "ID of the antonym",
                        "type": "integer"
                    },
                    "word": {
                        "description": "Antonym word",
                        "type": "string"
                    },
                    "definition_id": {
                        "description": "ID of the related definition",
                        "type": "integer"
                    },
                    "definition": {
                        "$ref": "#/components/schemas/Definition"
                    }
                },
                "type": "object"
            },
            "DefinitionSynonym": {
                "title": "DefinitionSynonym",
                "description": "Synonym for the definition",
                "properties": {
                    "id": {
                        "description": "ID of the synonym",
                        "type": "integer"
                    },
                    "word": {
                        "description": "Synonym word",
                        "type": "string"
                    },
                    "definition_id": {
                        "description": "ID of the related definition",
                        "type": "integer"
                    },
                    "definition": {
                        "$ref": "#/components/schemas/Definition"
                    }
                },
                "type": "object"
            },
            "Meaning": {
                "title": "Meaning",
                "required": [
                    "part_of_speech",
                    "word_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the meaning",
                        "type": "integer"
                    },
                    "part_of_speech": {
                        "description": "Part of speech",
                        "type": "string"
                    },
                    "word_id": {
                        "description": "ID of the word related to this meaning",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "description": "Deletion timestamp",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "word": {
                        "$ref": "#/components/schemas/Word"
                    },
                    "definitions": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Definition"
                        }
                    },
                    "synonyms": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MeaningSynonym"
                        }
                    },
                    "antonyms": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/MeaningAntonym"
                        }
                    }
                },
                "type": "object"
            },
            "MeaningAntonym": {
                "title": "MeaningAntonym",
                "description": "Antonym for the meaning",
                "properties": {
                    "id": {
                        "description": "ID of the antonym",
                        "type": "integer"
                    },
                    "word": {
                        "description": "Antonym word",
                        "type": "string"
                    },
                    "meaning_id": {
                        "description": "ID of the related meaning",
                        "type": "integer"
                    },
                    "meaning": {
                        "$ref": "#/components/schemas/Meaning"
                    }
                },
                "type": "object"
            },
            "MeaningSynonym": {
                "title": "MeaningSynonym",
                "description": "Synonym for the meaning",
                "properties": {
                    "id": {
                        "description": "ID of the synonym",
                        "type": "integer"
                    },
                    "word": {
                        "description": "Synonym word",
                        "type": "string"
                    },
                    "meaning_id": {
                        "description": "ID of the related meaning",
                        "type": "integer"
                    },
                    "meaning": {
                        "$ref": "#/components/schemas/Meaning"
                    }
                },
                "type": "object"
            },
            "Phonetic": {
                "title": "Phonetic",
                "required": [
                    "text",
                    "audio",
                    "word_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the phonetic",
                        "type": "integer"
                    },
                    "text": {
                        "description": "Phonetic transcription",
                        "type": "string"
                    },
                    "audio": {
                        "description": "URL of the audio pronunciation",
                        "type": "string"
                    },
                    "word_id": {
                        "description": "ID of the word related to this phonetic",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "description": "Deletion timestamp",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "word": {
                        "$ref": "#/components/schemas/Word"
                    }
                },
                "type": "object"
            },
            "SourceUrl": {
                "title": "Source URL",
                "required": [
                    "url",
                    "word_id"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the source URL",
                        "type": "integer"
                    },
                    "url": {
                        "description": "URL source",
                        "type": "string"
                    },
                    "word_id": {
                        "description": "ID of the word related to this source",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "description": "Deletion timestamp",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "word": {
                        "$ref": "#/components/schemas/Word"
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User",
                "description": "User model",
                "required": [
                    "id",
                    "name",
                    "email"
                ],
                "properties": {
                    "id": {
                        "description": "User's unique identifier",
                        "type": "string",
                        "format": "uuid",
                        "example": "9d43789a-1553-470c-b0ee-b33e000139bb"
                    },
                    "name": {
                        "description": "User's name",
                        "type": "string",
                        "example": "User 1"
                    },
                    "email": {
                        "description": "User's email address",
                        "type": "string",
                        "format": "email",
                        "example": "example@email.com"
                    },
                    "created_at": {
                        "description": "Timestamp when the user was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-17T02:11:57.000000Z"
                    },
                    "updated_at": {
                        "description": "Timestamp when the user was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-17T02:11:57.000000Z"
                    },
                    "history": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Word"
                        }
                    },
                    "favorites": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Word"
                        }
                    }
                },
                "type": "object"
            },
            "Word": {
                "title": "Word",
                "required": [
                    "word",
                    "license",
                    "license_url"
                ],
                "properties": {
                    "id": {
                        "description": "ID of the word",
                        "type": "integer"
                    },
                    "word": {
                        "description": "The word",
                        "type": "string"
                    },
                    "license": {
                        "description": "License of the word",
                        "type": "string"
                    },
                    "license_url": {
                        "description": "License URL of the word",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "description": "Deletion timestamp",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "phonetics": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Phonetic"
                        }
                    },
                    "meanings": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Meaning"
                        }
                    },
                    "sourceUrls": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SourceUrl"
                        }
                    },
                    "viewer": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    },
                    "users": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
}